fix: rename node/cotentLoader.ts to node/contentLoader.ts
