+++ src/node/index.ts
@@ -4,7 +4,7 @@ export * from './markdown'
 export * from './build/build'
 export * from './serve/serve'
 export * from './init/init'
-export * from './cotentLoader'
+export * from './contentLoader'
 export { defineLoader, type LoaderModule } from './plugins/staticDataPlugin'
 export { loadEnv } from 'vite'
